test: reject consent when test
